requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
itsdangerous==1.1.0
pathspec==0.8.0
pycparser==2.20
python-telegram-bot==12.7
pytz==2020.1
regex==2020.4.4
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
tornado==6.0.4
typed-ast==1.4.1
tzlocal==2.0.0
coloredlogs==14.0
humanfriendly==8.2
